support sub-tables with nested routes & jotai scoped providers
