+++ src/sources/TableSourceFirestore/index.ts
@@ -1,0 +1,2 @@
+export * from "./TableSourceFirestore";
+export { default } from "./TableSourceFirestore";
